@@ -26,7 +26,7 @@
     <parent>
         <artifactId>archetype</artifactId>
         <groupId>it.cnr.si</groupId>
-        <version>1.0.15</version>
+        <version>1.2.0</version>
     </parent>
     <modules>
         <module>siopelpus-commons</module>
@@ -35,9 +35,10 @@
     </modules>
     <properties>
         <sonar.skip>true</sonar.skip>
-        <java.version>1.8</java.version>
-        <spring.version>5.2.5.RELEASE</spring.version>
+        <java.version>21</java.version>
+        <spring.version>7.0.2</spring.version>
         <slf4j.version>1.7.21</slf4j.version>
+        <jakarta.xml.bind.version>4.0.2</jakarta.xml.bind.version>
     </properties>
     <build>
         <pluginManagement>
