@@ -25,18 +25,20 @@
 
 package it.cnr.si.siopeplus;
 
+import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.List;
+
 import it.siopeplus.CtEsitoMandati;
 import it.siopeplus.CtEsitoReversali;
 import it.siopeplus.CtTestataMessaggio;
 import it.siopeplus.SignatureType;
-import org.jvnet.jaxb2_commons.lang.JAXBToStringStrategy;
-import org.jvnet.jaxb2_commons.lang.ToString2;
-import org.jvnet.jaxb2_commons.lang.ToStringStrategy2;
-import org.jvnet.jaxb2_commons.locator.ObjectLocator;
-
-import javax.xml.bind.annotation.*;
-import java.util.ArrayList;
-import java.util.List;
+import jakarta.xml.bind.JAXBElement;
+import jakarta.xml.bind.annotation.*;
+import org.jvnet.jaxb.lang.JAXBToStringStrategy;
+import org.jvnet.jaxb.lang.ToString;
+import org.jvnet.jaxb.lang.ToStringStrategy;
+import org.jvnet.jaxb.locator.ObjectLocator;
 
 
 /**
@@ -73,7 +75,7 @@ import java.util.List;
     "signature"
 })
 @XmlRootElement(name = "messaggi_esito_applicativo")
-public class MessaggiEsitoApplicativo implements ToString2
+public class MessaggiEsitoApplicativo implements Serializable, ToString
 {
 
     @XmlElement(name = "testata_messaggio", required = true)
@@ -217,21 +219,25 @@ public class MessaggiEsitoApplicativo implements ToString2
         this.id = value;
     }
 
+    @Override
     public String toString() {
-        final ToStringStrategy2 strategy = JAXBToStringStrategy.INSTANCE;
+        final ToStringStrategy strategy = JAXBToStringStrategy.getInstance();
         final StringBuilder buffer = new StringBuilder();
         append(null, buffer, strategy);
         return buffer.toString();
     }
 
-    public StringBuilder append(ObjectLocator locator, StringBuilder buffer, ToStringStrategy2 strategy) {
+    @Override
+    public StringBuilder append(ObjectLocator locator, StringBuilder buffer, ToStringStrategy strategy) {
         strategy.appendStart(locator, this, buffer);
         appendFields(locator, buffer, strategy);
         strategy.appendEnd(locator, this, buffer);
         return buffer;
     }
 
-    public StringBuilder appendFields(ObjectLocator locator, StringBuilder buffer, ToStringStrategy2 strategy) {
+    @Override
+    public StringBuilder appendFields(ObjectLocator locator, StringBuilder buffer, ToStringStrategy strategy) {
+
         {
             CtTestataMessaggio theTestataMessaggio;
             theTestataMessaggio = this.getTestataMessaggio();
