@@ -30,9 +30,9 @@ import it.cnr.si.siopeplus.MessaggioRicezioneFlusso;
 import it.cnr.si.siopeplus.MessaggioRifiutoFlusso;
 import it.siopeplus.*;
 
-import javax.xml.bind.JAXBElement;
-import javax.xml.bind.annotation.XmlElementDecl;
-import javax.xml.bind.annotation.XmlRegistry;
+import jakarta.xml.bind.JAXBElement;
+import jakarta.xml.bind.annotation.XmlElementDecl;
+import jakarta.xml.bind.annotation.XmlRegistry;
 import javax.xml.namespace.QName;
 
 
